LogicalButton
