+++ Source/Frontend/UI/UI.csproj
@@ -381,6 +381,7 @@
     <Compile Include="Input\Input.cs" />
     <Compile Include="Input\IPCKeyInput.cs" />
     <Compile Include="Input\Keyboard.cs" />
+    <Compile Include="Input\LogicalButton.cs" />
     <Compile Include="Input\WorkingDictionary.cs" />
     <Compile Include="Modular\UI_CanvasForm.cs">
       <SubType>Form</SubType>
